refactor: isolate terminal shell dialect renderers
